feat(frontend): apply animated food grid background to every page (#772)
