+++ app/frontend/src/components/AppBackground.css
@@ -1,0 +1,13 @@
+.app-bg {
+  position: fixed;
+  inset: 0;
+  z-index: 0;
+}
+
+.app-bg-filter {
+  position: fixed;
+  inset: 0;
+  z-index: 1;
+  background: rgba(0, 0, 0, 0.7);
+  pointer-events: none;
+}
