+++ app/frontend/src/styles/global.css
@@ -78,6 +78,8 @@ a:hover {
   max-width: 960px;
   margin: 0 auto;
   padding: 2.5rem 1.5rem 5rem;
+  position: relative;
+  z-index: 2;
 }
 
 .page-card {
